rag-system/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
python-multipart
httpx
PyPDF2
python-dotenv
pydantic-settings
numpy
